+++ game_core/edit_mode/tilesets_tab/brush_manager.py
