+++ game_core/edit_mode/tilesets_tab/layer_manager.py
