+++ game_core/edit_mode/tilesets_tab/tileset_components/__init__.py
@@ -1,0 +1,1 @@
+# Placeholder for __init__.py
