+++ game_core/edit_mode/tilesets_tab/tileset_components/character_tilesets.py
@@ -1,0 +1,1 @@
+# Placeholder for character_tilesets.py
