+++ game_core/edit_mode/tilesets_tab/tileset_components/dungeon_anim_tilesets.py
@@ -1,0 +1,1 @@
+# Placeholder for dungeon_anim_tilesets.py
