+++ game_core/edit_mode/tilesets_tab/tileset_components/enemies_tilesets.py
@@ -1,0 +1,1 @@
+# Placeholder for enemies_tilesets.py
