+++ game_core/edit_mode/tilesets_tab/tileset_components/overworld_tilesets.py
@@ -1,0 +1,1 @@
+# Placeholder for overworld_tilesets.py
